Create separate svgo transformer. Update logic of svg-react transformer.
