+++ packages/transformers/svgo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@parcel/transformer-svg-react",
+  "name": "@parcel/transformer-svgo",
   "version": "2.0.0-alpha.3.2",
   "license": "MIT",
   "publishConfig": {
@@ -9,15 +9,13 @@
     "type": "git",
     "url": "https://github.com/parcel-bundler/parcel.git"
   },
-  "main": "lib/SvgReactTransformer.js",
-  "source": "src/SvgReactTransformer.js",
+  "main": "lib/SVGOTransformer.js",
+  "source": "src/SVGOTransformer.js",
   "engines": {
     "parcel": "^2.0.0-alpha.1.1"
   },
   "dependencies": {
-    "@babel/parser": "^7.8.4",
     "@parcel/plugin": "^2.0.0-alpha.3",
-    "camelcase": "^5.3.1",
-    "svg-to-jsx": "^1.0.2"
+    "svgo": "^1.3.2"
   }
 }
